не исключены отрицательные значения в FoodList.js исправлено отставание стейта не убран старый список при поиске

+++ src/components/food-list/FoodList.js
@@ -1,54 +1,30 @@
 import "./FoodList.css";
-import {useEffect, useState} from "react";
-import {stashDataSession, stashDataStorage} from "../../services/localStorageDB";
+import {useEffect} from "react";
+import {stashDataSession} from "../../services/localStorageDB";
 import ListElement from "../list-item/ListItem";
 import AddNewItem from "../add-new-item/AddNewItem";
+import SearchProducts from "../search-product/SearchProducts";
 
-const FoodList = ({data}) => {
-	const [productList, setProductList] = useState(JSON
-		.parse(sessionStorage.getItem("products") || localStorage.getItem("products")) || data);
-	const [amount, setAmount] = useState(0);
+const FoodList = ({
+	                  changeAmount,
+	                  amount,
+	                  setProductList,
+	                  setAmount,
+	                  productList,
+	                  clearAmount,
+	                  deleteItemFromStorage,
+	                  deleteItem
+                  }) => {
 	
-	const changeAmount = (e, property) => {
-		setAmount(amount => (parseInt(e.target.value)));
-		let indexOfChangedItem = productList.findIndex(item => item.name === e.target.name);
-		const newItem = [...productList].filter(item => item.name === e.target.name);
-		newItem[0] = {...newItem[0], [property]: e.target.value};
-		
-		const oldItemsBefore = [...productList].filter((item, i) => i < indexOfChangedItem);
-		const oldItemsAfter = [...productList].filter((item, i) => i > indexOfChangedItem);
-		
-		setProductList(list => [...oldItemsBefore, newItem[0], ...oldItemsAfter]);
-		stashDataSession("products", productList);
-	};
-	
-	const clearAmount = (arr) => {
-		setProductList(() => arr.map(item => ({...item, amount: 0})));
-		setAmount(0);
-		stashDataSession("products", productList);
-	};
-	
-	const deleteItem = (e) => {
-		const newArr = productList.filter(item => item.name !== e.target.name);
-		setProductList(productList => newArr);
-		stashDataSession("products", newArr);
-	};
-	
-	const deleteItemFromStorage = (e) => {
-		const newArr = productList.filter(item => item.name !== e.target.name);
-		setProductList(productList => newArr);
-		stashDataSession("products", newArr);
-		stashDataStorage("products", newArr);
-	};
-	
+	//
 	useEffect(() => {
 		stashDataSession("products", productList);
 	}, []);
-	
+
 	useEffect(() => {
 		stashDataSession("products", productList);
 	}, [productList]);
-	
+
 	useEffect(() => {
 		setProductList(list => JSON.parse(sessionStorage.getItem("products")))
 	}, [amount]);
@@ -57,28 +33,43 @@ const FoodList = ({data}) => {
 		return Math.ceil(price * (amount / pack));
 	};
 	
-	const list = productList.map((item, index) => {
-		return <ListElement key={index}
-		                    changeAmount={changeAmount}
-		                    item={item} index={index}
-		                    calulatePriceOfProduct={calculatePriceOfProduct}
-		                    deleteItem={deleteItem}
-		                    deleteItemFromStorage={deleteItemFromStorage}
-		                    />;
-	});
+	const listFormation = () => {
+		let result;
+		if (sessionStorage.getItem("products")) {
+			result = JSON.parse(sessionStorage.getItem("products")).map((item, index) => {
+				return <ListElement key={index}
+				                    changeAmount={changeAmount}
+				                    item={item} index={index}
+				                    calulatePriceOfProduct={calculatePriceOfProduct}
+				                    deleteItem={deleteItem}
+				                    deleteItemFromStorage={deleteItemFromStorage}
+				/>;
+			});
+		} else {
+			result = <li>'Нет выбранных продуктов'</li>;
+		}
+		return result;
+	};
+	
+	console.log('parced productlist', productList);
+	const list = listFormation();
+	const finalPrice = productList.reduce((a, b) => a + calculatePriceOfProduct(b.price, b.amount, b.pack), 0);
+	// useEffect(() => {
+	// 	changeAmount()
+	// }, [amount, productList])
 	
 	return (
 		<>
 			<AddNewItem setProductList={setProductList} productList={productList}/>
 			<div className="product_list">
+				<SearchProducts setProductList={setProductList} productList={productList}/>
 				<h2 className="product_list__heading">Список продуктов</h2>
 				<ul className="product_list__elements">
 					{list}
 				</ul>
 				<div>
 					<h3>Общая стоимость использованных продуктов:<span className="span">_</span> <u>
-						{productList
-							.reduce((a, b) => a + calculatePriceOfProduct(b.price, b.amount, b.pack), 0)}
+						{finalPrice}
 					</u>
 					</h3>
 					<button>Сохранить стоимость ингредиентов</button>
